el punto 8

+++ Practicas/Practica3/ruido_gaussiano.m
@@ -3,5 +3,5 @@ function ruido = ruido_gaussiano(img, sigma, mu)
 img = double(img)/255;
 gen = mu + (sigma-mu)*randn(size(img));
 ruido = gen + img;
-ruido = uint8(ruido*255);
+ruido = (ruido*255);
 endfunction
